In OracleEvaluator, return 0 if getInitialValue is out of boundary.
OracleKTest knows the exact size of oracle input. However during replay,
we may init symbolic objects with size larger than oracle input. If we
init symbolic objects with wrong size, the constraints may contain out
of boundary access. We evalue these wrong accesses to 0, which will not
affect correctness.

The root cause of this issue is that klee does not support malloc with
symbolic size (in turn cannot support file with symbolic size, file
buffer need to be preallocated).