WORKAROUND, assignment returned by solver could lead to out of bound access
When we are dealing with bugs, it is possible that some array indices
are not constrainted properly (which is actually a bug).
So, assignment returned by solver could lead to out of bound access when
evaluating constraints.
If that happens, constraints may not be reduced to ConstantExpr when
applying assignment, which is considered as error previously.

I do not think carefully about this issue yet, but current workaround is
to report those "still-symbolic" constraints evaluation instead of aborting.