Update gephi hase python script. Consider ConstantExpr concretizable directly.
We want to find and color concretizable nodes after we select a few
nodes to record. Previously we made aggressive simplification to the
graph to only show the dependency. Now we need to also consider kinst
(recordable or not), so we may no longer do such simplifications. As a
result, we need to expand our definition of "concretizable".
We used to only consider nodes fully depend on recorded nodes are
"concretizable".