Do not deserialize .path_datarec if it is empty
This patch is necessary since currently .path_datarec will always be
generated together with the .path file.
Note that the fstream will be good() when you open a file, even if that file
is empty.
If you peek to see its first character, the eof will be triggered if the
file is empty.