In GraphvizDrawer, also assign Width attribute to UpdateNode
There is no "Width" member in class UpdateNode. However, we need Width
of all nodes in the constraint graph to compute recording priority heuristics.
Now, we make the same assumption as KLEE's developer: UpdateNode will
only represent a single byte update. So the Width will be set to
a constant 8.