Do not overwrite existing kinst if we are not in the target program
Since I prefer to not instrument external library (e.g. POSIX runtime
and LIBC), I need to take "kinst overwritting" into consideration.
Note that when you pass the same symbolic expression around, previous
behavior is to keep updating (overwritting the kinst field of that
symbolic expression).
Now, we mark LLVM functions from POSIX and LIBC with special function
attributes. Then we only bind a kinst to a symbolic expression in either
of the following scenarios:
1) the symbolic expression has not been bound to any kinst
2) the kinst is from the target program (not from POSIX nor LIBC)

TODO: the policy of when should we update/overwrite the kinst of a
symbolic expression can be flexible.
e.g. you can take the instruction frequency into consideration.
There is a tradeoff: consider the following call stack: (lower is caller
higher is callee)
|-----------------------  call stack end  ---------------------------------|
|                        high frequency func6: reg R6                      |
|             symbolic func4: reg R4              | concrete func5: reg R5 |
| symbolic func1: reg R1 | symbolic func2: reg R2 | concrete func3: reg R3 |
|                        high frequency func0: reg R0                      |
|-----------------------  call stack start  -------------------------------|
a local register/stack var R0 in the func0 is passed to func1 or func2
or func3. In each function, the value of R0 is aliased to R1, R2, R3
respectively. Assuming func1 and func2 are symbolic (we want to record
more to concretize them) and func3 is already concrete. The rest of the
call stack can be interpreted similarly. The same symbolic expression is
bound to R0...R6 during replay.

Now let us consider the problem of which register should we record.
During a incomplete symbolic replay, we may only encounter
func0->func1->func4->func6.
If we chose to record:
1) R0: many recorded value may be concrete during replay, thus useless
2) R1: recorded value will be useful but you need to stop again when
you encounter func2. That leads to more iterations (require bugs
reoccuring more times).
3) R4: the ideal policy.Bbut how can you tell func2 is symbolic or not
just based on a partial replay.
4) R6: the previous behavior, always overwritting, bind to the last instruction.
These choice can be achieved by using different kinst overwritting
policy.