Add a socket simulator to POSIX runtime.
The socker simulator consists of a socket handler base class.
You can derive from it and define your own handler.
For now, it supports hooks in "bind()" and "listen()", which should be
enough to replay the memcached bug.

Besides, a new cmdline option "-symbolic-sock-handler" can inform if all
socket simulator should behave like during a symbolic replay (similar to
"-sym-file").

I also implement a simulator to reproduce bugs in memcached1.5.13. Note
that the listen handler creates a new thread to "connect" then "write"
to interesting server sockets.
That is because "connect" and "accept" are both blocking and could not happen in
the same thread.