Parse "joinable" thread option from pthread_attr_t during pthread_create.
This was not implemented in cloud9 (marked as TODO). In cloud9, all
threads created are joinable, thus occupying a valid thread index even
if a thread exits. I think fast reclaim of thread index is necessary so
I implemented this.
Note that there are many attributes in "pthread_attr_t, but currently I
only take care of the joinable.