[Gephi] analyze all recordable instructions and add a new heuristic "remainScore"
It has been issue for a while that when can you skip nodes in the
constraint graph when analyzing recordable instructions.
Previously, I skipped any kinst A if it can be concretized by another
kinst B. This approach ignored that an intermediate kinst A could concretize
a few less nodes with significant less data recorded than kinst B.
Without having some smart algorithm to decide when to skip, now I do not
skip kinst at all.

A new heuristic "remainScore":
Existing heuristics only focus on properties of eliminated nodes on the
given constraint graph.
The "remainScore" instead focuses on properties of remaining nodes
(cannot be concretized) on the constraint graph assuming a certain kinst
is recorded. I write this for the matrixssl bug, in which "remainScore"
highlights some recording candidates significantly shorten the indirect
depth but only concretize a tiny fraction of the constraint graph.

MISC:
1. Print if each recordable instruction represents a pointer or not.
2. Move heuristics and properties calculation of each recordable
   instruction to class constructor. This is for reducing repetitive
   computation when accessing those heuristics/properties later.
   Especially, the new "remainScore" requies expensive subgraph
   traversal.