Dump Ambiguous memory address kquery before doing the multi-resolution solver call
Since the multi-resolution solver call may take a lot of time, I should
dump the constraint earlier and analyze it without the multi-resolution
result.
I want to the multi-resolution result because you can only assert an
address is out of bound if it is proven to have no resolution.
In practice, I tend to believe a real out-of-bound address is concrete
and always resolved quickly.
While a multi-resolution address will spend a lot of time in solver.