[Gephi] Add a new option to count length of each update list upon all symbolic arraies.
The new heuristic, which measures how many huge arraies are going to be
passed to the solver, should be aware of arraies initially not big, but
having a super long update list as well.

Note that this commit only focuses on "update list". You should refer to
the origin kquery to find out huge arraies which are not updated (no
UpdateNode), but still read by some expression thus still exist in the
constraints.

Note that this option should work with NON-simplified cosntraint graph.
Because "simplify-drawing" will fold chains of concrete update
(UpdateNode).