[Expr simplification] I should preserve kinst information as much as possible
KLEE will do various expression transformation and most of them
currently destroy the kinst information.
I need to look into each transformation and preserve the kinst as long
as the transformation result can still be concretized when I record the
kinst of the origin expression.

This commit keeps the kinst after double-negation elimination.