[Gephi] Change Node concretization algorithm to focus on a set nodes (instead of a set of instructions)
Previously, the API was taking a single node, return a set of
instruction identifiers (kinst) representing the best recording strategy.
And the iterative optimization process also replace an instruction X with
others Y only if all nodes bound to X can be concretized by Y.

Now, we instead return the set of nodes as the best recording strategy
for a single node. And during the iterative optimization process, we
will replace an instruction X with others Y if all nodes bound to X AND
locate in current recording strategy can be concretized by Y.
For nodes also bound to X but not in current recording strategy, I do not care.