[Expr] Make equivalency cache last longer by delay deallocation.
In ExprReplaceVisitor{Single, Multi}, the replaced Expr/UpdateNode could
be deallocated, in which case the equivalency caches (Expr::equivs and
UpdateNode::UNequivs) have to be flushed.
Because of performance concern, I want to delay deallocation via putting
replaced Expr/UpdateNode to a temporary buffer until the entier
replacement process finishes.

To unify the "delay deallocation and cache equivalency longer"
behaviour, I added a new function call "replace()" to
ExprReplaceVisitorBase which serves as the main entrance call
(previously is visit()).