[POSIX] Add /dev/urandom modeling
Model /dev/urandom as a special concrete or (not pure) symbolic file.
User can either assume its content is symbolic but size and stats are
known. Or, user can use the content of a concrete file as /dev/urandom.
They are readonly and their read offsets do not change when re-opened/closed.