[Gephi Py] Respect the "ALLOWPTR" option even the given initial recording strategy contains pointers.
There is an "ALLOWPTR" option to decide if recording pointer directly is
allowed. When this option is disabled, you may still start with pointer
recording. E.g. during Ambiguous Memory addresses, the expressions
needed to be concretized are always pointers, which composes the initial
recording strategy.

This commit handles the case that give an initial recording strategy
which records pointers, the refined recording strategy should record
non-pointer values if there is any, even if non-pointer values are more
expensive than pointers.