[Expr] Use std::unordered_set instead of std::vector as the container of a collection of constraints
This commit changes the definition of type `Constraints_ty`, which
generally represents a collection of constraints.
Such change is made to accommodate ConstraintManager constraints
management optimization (which is not included in this commit).
In future commits, I want to avoid scanning the entire constraints
vector when adding/deleting a single constraint.