[WIP] ugly patch to work with llvm14
This is ugly because I do not group related changes together and I do not port klee commits here.
Instead, I mostly rely on compiler errors to come up with this commit.

It is better if I can seperate this patch according to the upstream klee.
And for the best case, I should upstream my changes.